requirements.txt
gevent
Adafruit_DHT
flask
requests
RPi.GPIO
Flask-SQLAlchemy